requirements.txt
openai>=1.0.0
dspy-ai>=2.4.0